requirements.txt
google-genai==0.3.0
numpy==1.26.4
streamlit
tensorflow
tf-keras
pillow